Separate out winmd from winrt; retain WinRT string helpers
