+++ lib/src/winrt/data/json/ijsonarraystatics.dart
@@ -14,7 +14,6 @@ import '../../../macros.dart';
 import '../../../types.dart';
 import '../../../utils.dart';
 import '../../../winrt/data/json/jsonarray.dart';
-import '../../../winrt_helpers.dart';
 
 /// @nodoc
 const IID_IJsonArrayStatics = '{DB1434A9-E164-499F-93E2-8A8F49BB90BA}';
