+++ lib/src/winrt/data/json/ijsonobjectstatics.dart
@@ -14,7 +14,6 @@ import '../../../macros.dart';
 import '../../../types.dart';
 import '../../../utils.dart';
 import '../../../winrt/data/json/jsonobject.dart';
-import '../../../winrt_helpers.dart';
 
 /// @nodoc
 const IID_IJsonObjectStatics = '{2289F159-54DE-45D8-ABCC-22603FA066A0}';
